Replay: fixed FMT messages in replay output
